+++ lib/pure/os.nim
@@ -1550,7 +1550,10 @@ proc sameFile*(path1, path2: string): bool {.rtl, extern: "nos$1",
     else:
       result = a.st_dev == b.st_dev and a.st_ino == b.st_ino
 
+<<<<<<< HEAD
 
+=======
+>>>>>>> 33d79b9e64f8f67adcf8415de61d1568630e36cb
 type
   FilePermission* = enum   ## File access permission, modelled after UNIX.
     ##
@@ -3041,6 +3044,8 @@ type
     lastAccessTime*: times.Time       ## Time file was last accessed.
     lastWriteTime*: times.Time        ## Time file was last modified/written to.
     creationTime*: times.Time         ## Time file was created. Not supported on all systems!
+    blockSize*: int                   ## Preferred I/O block size for this object.
+                                      ## In some filesystems, this may vary from file to file.
 
 template rawToFormalFileInfo(rawInfo, path, formalInfo): untyped =
   ## Transforms the native file info structure into the one nim uses.
@@ -3055,6 +3060,7 @@ template rawToFormalFileInfo(rawInfo, path, formalInfo): untyped =
     formalInfo.lastAccessTime = fromWinTime(rdFileTime(rawInfo.ftLastAccessTime))
     formalInfo.lastWriteTime = fromWinTime(rdFileTime(rawInfo.ftLastWriteTime))
     formalInfo.creationTime = fromWinTime(rdFileTime(rawInfo.ftCreationTime))
+    formalInfo.blockSize = 8192 # xxx use windows API instead of hardcoding
 
     # Retrieve basic permissions
     if (rawInfo.dwFileAttributes and FILE_ATTRIBUTE_READONLY) != 0'i32:
@@ -3081,6 +3087,7 @@ template rawToFormalFileInfo(rawInfo, path, formalInfo): untyped =
     formalInfo.lastAccessTime = rawInfo.st_atim.toTime
     formalInfo.lastWriteTime = rawInfo.st_mtim.toTime
     formalInfo.creationTime = rawInfo.st_ctim.toTime
+    formalInfo.blockSize = rawInfo.st_blksize
 
     formalInfo.permissions = {}
     checkAndIncludeMode(S_IRUSR, fpUserRead)
@@ -3184,7 +3191,7 @@ proc getFileInfo*(path: string, followSymlink = true): FileInfo {.noWeirdTarget.
         raiseOSError(osLastError(), path)
     rawToFormalFileInfo(rawInfo, path, result)
 
-proc sameFileContent*(path1, path2: string; checkSize = false; checkFiles = false; bufferSize = 8192.Positive): bool {.
+proc sameFileContent*(path1, path2: string; checkSize = false; checkFiles = false): bool {.
     rtl, extern: "nos$1", tags: [ReadIOEffect], noWeirdTarget.} =
   ## Returns `true` if both pathname arguments refer to files with identical binary content.
   ##
@@ -3194,14 +3201,13 @@ proc sameFileContent*(path1, path2: string; checkSize = false; checkFiles = fals
   ## This does not fail if the file never existed in the first place, unless `checkFiles = true`.
   ##
   ## .. code-block:: nim
-  ##   echo sameFileContent("file0.txt", "file1.txt", checkSize = true, bufferSize = 999)
+  ##   echo sameFileContent("file0.txt", "file1.txt", checkSize = true)
   ##
   ## See also:
   ## * `sameFile proc <#sameFile,string,string>`_
   var a, b: File
   var mustRead = true
-  var bufA = alloc(bufferSize)
-  var bufB = alloc(bufferSize)
+  var bufA, bufB: pointer
   try:  # readBuffer or open may or may not raise IOError.
     if not open(a, path1):
       if checkFiles:
@@ -3213,7 +3219,12 @@ proc sameFileContent*(path1, path2: string; checkSize = false; checkFiles = fals
         raise newException(IOError, "Can not open file: $1" % [path2])
       else:
         mustRead = false
-    if checkSize and getFileInfo(a).size != getFileInfo(b).size: mustRead = false
+    let aInfo = getFileInfo(a)
+    let bInfo = getFileInfo(b)
+    let bufferSize = max(aInfo.blockSize, bInfo.blockSize)
+    bufA = alloc(bufferSize)
+    bufB = alloc(bufferSize)
+    if checkSize and aInfo.size != bInfo.size: mustRead = false
     if mustRead:
       while true:
         var readA = readBuffer(a, bufA, bufferSize)